third_party/blink/web_tests/svg/dynamic-updates/SVGFEDiffuseLightingElement-dom-diffuseConstant-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'diffuseConstant' attribute of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.getAttribute('diffuseConstant') is "0"
PASS gradientElement.getAttribute('diffuseConstant') is "1"
PASS successfullyParsed is true

TEST COMPLETE
